+++ cargo-rbmt/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Add `progress` mode to `RBMT_LOG_LEVEL` for interactive use.
 * Add trailing newline clean up to `fmt`.
+* More robust lockfile management when using the baseline feature of `test`.
 
 ## [0.1.0] - 2026-03-20